requirements.txt
# MCP Server dependencies
fastmcp>=0.1.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database support
pymysql>=1.1.0
cryptography>=42.0.0
